src/main/java/com/psygate/datastructures/spatial/D2Point.java
/*
 * General datastructures.
 * Copyright (C) 2015  psygate (https://github.com/psygate)
 * 
 * This program is free software: you can redistribute it and/or modify
 * it under the terms of the GNU General Public License as published by
 * the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version.
 * 
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU General Public License for more details.
 * 
 * You should have received a copy of the GNU General Public License
 * along with this program.  If not, see <https://www.gnu.org/licenses/gpl-3.0.html>.
 * 
 */
package com.psygate.datastructures.spatial;

import com.psygate.datastructures.spatial.ID2Point;

/**
 * Implementation of the IDPoint interface.
 *
 * @see ID2Point
 * @author psygate (https://github.com/psygate)
 */
public final class D2Point implements ID2Point {

    private final double x, y;
    
    /**
     *
     * @param x X coordinate of the new point.
     * @param y Y coordinate of the new point.
     */
    public D2Point(final double x, final double y) {
        this.x = x;
        this.y = y;
    }

    /**
     *
     * @param point Point to copy, so that point.x == this.x &amp;&amp; point.y == this.y
     */
    public D2Point(final ID2Point point) {
        this.x = point.getX();
        this.y = point.getY();
    }

    @Override
    public int hashCode() {
        int hash = 5;
        hash = 59 * hash + (int) (Double.doubleToLongBits(this.x) ^ (Double.doubleToLongBits(this.x) >>> 32));
        hash = 59 * hash + (int) (Double.doubleToLongBits(this.y) ^ (Double.doubleToLongBits(this.y) >>> 32));
        return hash;
    }

    @Override
    public boolean equals(Object obj) {
        if (this == obj) {
            return true;
        }
        if (obj == null) {
            return false;
        }
        if (getClass() != obj.getClass()) {
            return false;
        }
        final D2Point other = (D2Point) obj;
        if (Double.doubleToLongBits(this.x) != Double.doubleToLongBits(other.x)) {
            return false;
        }
        if (Double.doubleToLongBits(this.y) != Double.doubleToLongBits(other.y)) {
            return false;
        }
        return true;
    }

    @Override
    public double getX() {
        return x;
    }

    @Override
    public double getY() {
        return y;
    }

    @Override
    public String toString() {
        return "DPoint{" + "x=" + x + ", y=" + y + '}';
    }

}
